feat: introduce GreatVibes class for file creation and CSV conversion
Refactor SqlServer_Helper class to make methods static and encapsulate connection string. As part of this change, the GreatVibes class was also introduced for file creation and CSV conversion. Program.cs was updated as well to invoke the CreateFiles method of GreatVibes. This simplifies database query executions and improves the clarity of the data flow in the application.