feat(sql): adjust database connection handling and data insertion methods
This commit modifies the `SqlServer_Helper` class enhancing database connection functionality with a fallback option and implements methods for data insertion into the database. An ErrorHandler was introduced for accurate error logging to a text file. A test project `GreatVinesTest` was added to the solution for unit testing ensuring the robustness of our ETL process and new packages were included for effective configuration management and effortless testing.